bigquery: add Client.load_table_from_file
Move the method from Table to Client.
