+++ bigquery/google/cloud/bigquery/table.py
@@ -17,13 +17,8 @@
 from __future__ import absolute_import
 
 import datetime
-import os
 
 import six
-
-from google import resumable_media
-from google.resumable_media.requests import MultipartUpload
-from google.resumable_media.requests import ResumableUpload
 
 from google.cloud import exceptions
 from google.cloud._helpers import _datetime_from_microseconds
@@ -34,17 +29,6 @@ from google.cloud.bigquery._helpers import _SCALAR_VALUE_TO_JSON_ROW
 
 _TABLE_HAS_NO_SCHEMA = "Table has no schema:  call 'client.get_table()'"
 _MARKER = object()
-_DEFAULT_CHUNKSIZE = 1048576  # 1024 * 1024 B = 1 MB
-_BASE_UPLOAD_TEMPLATE = (
-    u'https://www.googleapis.com/upload/bigquery/v2/projects/'
-    u'{project}/jobs?uploadType=')
-_MULTIPART_URL_TEMPLATE = _BASE_UPLOAD_TEMPLATE + u'multipart'
-_RESUMABLE_URL_TEMPLATE = _BASE_UPLOAD_TEMPLATE + u'resumable'
-_GENERIC_CONTENT_TYPE = u'*/*'
-_READ_LESS_THAN_SIZE = (
-    'Size {:d} was specified but the file-like object only had '
-    '{:d} bytes remaining.')
-_DEFAULT_NUM_RETRIES = 6
 
 
 class TableReference(object):
@@ -826,353 +810,6 @@ class Table(object):
 
         return errors
 
-    def _get_transport(self, client):
-        """Return the client's transport.
-
-        :type client: :class:`~google.cloud.bigquery.client.Client`
-        :param client: The client to use.
-
-        :rtype transport:
-            :class:`~google.auth.transport.requests.AuthorizedSession`
-        :returns: The transport (with credentials) that will
-                  make authenticated requests.
-        """
-        return client._http
-
-    def _initiate_resumable_upload(self, client, stream,
-                                   metadata, num_retries):
-        """Initiate a resumable upload.
-
-        :type client: :class:`~google.cloud.bigquery.client.Client`
-        :param client: The client to use.
-
-        :type stream: IO[bytes]
-        :param stream: A bytes IO object open for reading.
-
-        :type metadata: dict
-        :param metadata: The metadata associated with the upload.
-
-        :type num_retries: int
-        :param num_retries: Number of upload retries. (Deprecated: This
-                            argument will be removed in a future release.)
-
-        :rtype: tuple
-        :returns:
-            Pair of
-
-            * The :class:`~google.resumable_media.requests.ResumableUpload`
-              that was created
-            * The ``transport`` used to initiate the upload.
-        """
-        chunk_size = _DEFAULT_CHUNKSIZE
-        transport = self._get_transport(client)
-        headers = _get_upload_headers(client._connection.USER_AGENT)
-        upload_url = _RESUMABLE_URL_TEMPLATE.format(project=self.project)
-        upload = ResumableUpload(upload_url, chunk_size, headers=headers)
-
-        if num_retries is not None:
-            upload._retry_strategy = resumable_media.RetryStrategy(
-                max_retries=num_retries)
-
-        upload.initiate(
-            transport, stream, metadata, _GENERIC_CONTENT_TYPE,
-            stream_final=False)
-
-        return upload, transport
-
-    def _do_resumable_upload(self, client, stream, metadata, num_retries):
-        """Perform a resumable upload.
-
-        :type client: :class:`~google.cloud.bigquery.client.Client`
-        :param client: The client to use.
-
-        :type stream: IO[bytes]
-        :param stream: A bytes IO object open for reading.
-
-        :type metadata: dict
-        :param metadata: The metadata associated with the upload.
-
-        :type num_retries: int
-        :param num_retries: Number of upload retries. (Deprecated: This
-                            argument will be removed in a future release.)
-
-        :rtype: :class:`~requests.Response`
-        :returns: The "200 OK" response object returned after the final chunk
-                  is uploaded.
-        """
-        upload, transport = self._initiate_resumable_upload(
-            client, stream, metadata, num_retries)
-
-        while not upload.finished:
-            response = upload.transmit_next_chunk(transport)
-
-        return response
-
-    def _do_multipart_upload(self, client, stream, metadata,
-                             size, num_retries):
-        """Perform a multipart upload.
-
-        :type client: :class:`~google.cloud.bigquery.client.Client`
-        :param client: The client to use.
-
-        :type stream: IO[bytes]
-        :param stream: A bytes IO object open for reading.
-
-        :type metadata: dict
-        :param metadata: The metadata associated with the upload.
-
-        :type size: int
-        :param size: The number of bytes to be uploaded (which will be read
-                     from ``stream``). If not provided, the upload will be
-                     concluded once ``stream`` is exhausted (or :data:`None`).
-
-        :type num_retries: int
-        :param num_retries: Number of upload retries. (Deprecated: This
-                            argument will be removed in a future release.)
-
-        :rtype: :class:`~requests.Response`
-        :returns: The "200 OK" response object returned after the multipart
-                  upload request.
-        :raises: :exc:`ValueError` if the ``stream`` has fewer than ``size``
-                 bytes remaining.
-        """
-        data = stream.read(size)
-        if len(data) < size:
-            msg = _READ_LESS_THAN_SIZE.format(size, len(data))
-            raise ValueError(msg)
-
-        transport = self._get_transport(client)
-        headers = _get_upload_headers(client._connection.USER_AGENT)
-
-        upload_url = _MULTIPART_URL_TEMPLATE.format(project=self.project)
-        upload = MultipartUpload(upload_url, headers=headers)
-
-        if num_retries is not None:
-            upload._retry_strategy = resumable_media.RetryStrategy(
-                max_retries=num_retries)
-
-        response = upload.transmit(
-            transport, data, metadata, _GENERIC_CONTENT_TYPE)
-
-        return response
-
-    def _do_upload(self, client, stream, metadata, size, num_retries):
-        """Determine an upload strategy and then perform the upload.
-
-        If ``size`` is :data:`None`, then a resumable upload will be used,
-        otherwise the content and the metadata will be uploaded
-        in a single multipart upload request.
-
-        :type client: :class:`~google.cloud.bigquery.client.Client`
-        :param client: The client to use.
-
-        :type stream: IO[bytes]
-        :param stream: A bytes IO object open for reading.
-
-        :type metadata: dict
-        :param metadata: The metadata associated with the upload.
-
-        :type size: int
-        :param size: The number of bytes to be uploaded (which will be read
-                     from ``stream``). If not provided, the upload will be
-                     concluded once ``stream`` is exhausted (or :data:`None`).
-
-        :type num_retries: int
-        :param num_retries: Number of upload retries. (Deprecated: This
-                            argument will be removed in a future release.)
-
-        :rtype: dict
-        :returns: The parsed JSON from the "200 OK" response. This will be the
-                  **only** response in the multipart case and it will be the
-                  **final** response in the resumable case.
-        """
-        if size is None:
-            response = self._do_resumable_upload(
-                client, stream, metadata, num_retries)
-        else:
-            response = self._do_multipart_upload(
-                client, stream, metadata, size, num_retries)
-
-        return response.json()
-
-    # pylint: disable=too-many-arguments,too-many-locals
-    def upload_from_file(self,
-                         file_obj,
-                         source_format,
-                         rewind=False,
-                         size=None,
-                         num_retries=_DEFAULT_NUM_RETRIES,
-                         allow_jagged_rows=None,
-                         allow_quoted_newlines=None,
-                         create_disposition=None,
-                         encoding=None,
-                         field_delimiter=None,
-                         ignore_unknown_values=None,
-                         max_bad_records=None,
-                         quote_character=None,
-                         skip_leading_rows=None,
-                         write_disposition=None,
-                         client=None,
-                         job_name=None,
-                         null_marker=None):
-        """Upload the contents of this table from a file-like object.
-
-        :type file_obj: file
-        :param file_obj: A file handle opened in binary mode for reading.
-
-        :type source_format: str
-        :param source_format: Any supported format. The full list of supported
-            formats is documented under the
-            ``configuration.extract.destinationFormat`` property on this page:
-            https://cloud.google.com/bigquery/docs/reference/rest/v2/jobs
-
-        :type rewind: bool
-        :param rewind: If True, seek to the beginning of the file handle before
-                       writing the file.
-
-        :type size: int
-        :param size: The number of bytes to read from the file handle.
-                     If not provided, we'll try to guess the size using
-                     :func:`os.fstat`. (If the file handle is not from the
-                     filesystem this won't be possible.)
-
-        :type num_retries: int
-        :param num_retries: Number of upload retries. Defaults to 6.
-
-        :type allow_jagged_rows: bool
-        :param allow_jagged_rows: job configuration option;  see
-                                  :meth:`google.cloud.bigquery.job.LoadJob`.
-
-        :type allow_quoted_newlines: bool
-        :param allow_quoted_newlines: job configuration option; see
-                                      :meth:`google.cloud.bigquery.job.LoadJob`.
-
-        :type create_disposition: str
-        :param create_disposition: job configuration option; see
-                                   :meth:`google.cloud.bigquery.job.LoadJob`.
-
-        :type encoding: str
-        :param encoding: job configuration option; see
-                         :meth:`google.cloud.bigquery.job.LoadJob`.
-
-        :type field_delimiter: str
-        :param field_delimiter: job configuration option; see
-                                :meth:`google.cloud.bigquery.job.LoadJob`.
-
-        :type ignore_unknown_values: bool
-        :param ignore_unknown_values: job configuration option; see
-                                      :meth:`google.cloud.bigquery.job.LoadJob`.
-
-        :type max_bad_records: int
-        :param max_bad_records: job configuration option; see
-                                :meth:`google.cloud.bigquery.job.LoadJob`.
-
-        :type quote_character: str
-        :param quote_character: job configuration option; see
-                                :meth:`google.cloud.bigquery.job.LoadJob`.
-
-        :type skip_leading_rows: int
-        :param skip_leading_rows: job configuration option; see
-                                  :meth:`google.cloud.bigquery.job.LoadJob`.
-
-        :type write_disposition: str
-        :param write_disposition: job configuration option; see
-                                  :meth:`google.cloud.bigquery.job.LoadJob`.
-
-        :type client: :class:`~google.cloud.bigquery.client.Client`
-        :param client: (Optional) The client to use.  If not passed, falls back
-                       to the ``client`` stored on the current table.
-
-        :type job_name: str
-        :param job_name: Optional. The id of the job. Generated if not
-                         explicitly passed in.
-
-        :type null_marker: str
-        :param null_marker: Optional. A custom null marker (example: "\\N")
-
-        :rtype: :class:`~google.cloud.bigquery.jobs.LoadJob`
-
-        :returns: the job instance used to load the data (e.g., for
-                  querying status). Note that the job is already started:
-                  do not call ``job.begin()``.
-        :raises: :class:`ValueError` if ``size`` is not passed in and can not
-                 be determined, or if the ``file_obj`` can be detected to be
-                 a file opened in text mode.
-        """
-        client = self._require_client(client)
-        _maybe_rewind(file_obj, rewind=rewind)
-        _check_mode(file_obj)
-        metadata = _get_upload_metadata(
-            source_format, self._schema, self._project,
-            self._dataset_id, self.table_id)
-        _configure_job_metadata(metadata, allow_jagged_rows,
-                                allow_quoted_newlines, create_disposition,
-                                encoding, field_delimiter,
-                                ignore_unknown_values, max_bad_records,
-                                quote_character, skip_leading_rows,
-                                write_disposition, job_name, null_marker)
-
-        try:
-            created_json = self._do_upload(
-                client, file_obj, metadata, size, num_retries)
-            return client.job_from_resource(created_json)
-        except resumable_media.InvalidResponse as exc:
-            raise exceptions.from_http_response(exc.response)
-    # pylint: enable=too-many-arguments,too-many-locals
-
-
-def _configure_job_metadata(metadata,  # pylint: disable=too-many-arguments
-                            allow_jagged_rows,
-                            allow_quoted_newlines,
-                            create_disposition,
-                            encoding,
-                            field_delimiter,
-                            ignore_unknown_values,
-                            max_bad_records,
-                            quote_character,
-                            skip_leading_rows,
-                            write_disposition,
-                            job_name,
-                            null_marker):
-    """Helper for :meth:`Table.upload_from_file`."""
-    load_config = metadata['configuration']['load']
-
-    if allow_jagged_rows is not None:
-        load_config['allowJaggedRows'] = allow_jagged_rows
-
-    if allow_quoted_newlines is not None:
-        load_config['allowQuotedNewlines'] = allow_quoted_newlines
-
-    if create_disposition is not None:
-        load_config['createDisposition'] = create_disposition
-
-    if encoding is not None:
-        load_config['encoding'] = encoding
-
-    if field_delimiter is not None:
-        load_config['fieldDelimiter'] = field_delimiter
-
-    if ignore_unknown_values is not None:
-        load_config['ignoreUnknownValues'] = ignore_unknown_values
-
-    if max_bad_records is not None:
-        load_config['maxBadRecords'] = max_bad_records
-
-    if quote_character is not None:
-        load_config['quote'] = quote_character
-
-    if skip_leading_rows is not None:
-        load_config['skipLeadingRows'] = skip_leading_rows
-
-    if write_disposition is not None:
-        load_config['writeDisposition'] = write_disposition
-
-    if job_name is not None:
-        load_config['jobReference'] = {'jobId': job_name}
-
-    if null_marker is not None:
-        load_config['nullMarker'] = null_marker
-
 
 def _parse_schema_resource(info):
     """Parse a resource fragment into a schema field.
@@ -1220,53 +857,6 @@ def _build_schema_resource(fields):
         infos.append(info)
     return infos
 # pylint: enable=unused-argument
-
-
-def _maybe_rewind(stream, rewind=False):
-    """Rewind the stream if desired.
-
-    :type stream: IO[bytes]
-    :param stream: A bytes IO object open for reading.
-
-    :type rewind: bool
-    :param rewind: Indicates if we should seek to the beginning of the stream.
-    """
-    if rewind:
-        stream.seek(0, os.SEEK_SET)
-
-
-def _check_mode(stream):
-    """Check that a stream was opened in read-binary mode.
-
-    :type stream: IO[bytes]
-    :param stream: A bytes IO object open for reading.
-
-    :raises: :exc:`ValueError` if the ``stream.mode`` is a valid attribute
-             and is not among ``rb``, ``r+b`` or ``rb+``.
-    """
-    mode = getattr(stream, 'mode', None)
-
-    if mode is not None and mode not in ('rb', 'r+b', 'rb+'):
-        raise ValueError(
-            "Cannot upload files opened in text mode:  use "
-            "open(filename, mode='rb') or open(filename, mode='r+b')")
-
-
-def _get_upload_headers(user_agent):
-    """Get the headers for an upload request.
-
-    :type user_agent: str
-    :param user_agent: The user-agent for requests.
-
-    :rtype: dict
-    :returns: The headers to be used for the request.
-    """
-    return {
-        'Accept': 'application/json',
-        'Accept-Encoding': 'gzip, deflate',
-        'User-Agent': user_agent,
-        'content-type': 'application/json',
-    }
 
 
 def _get_upload_metadata(source_format, schema, project, dataset_id, table_id):
