+++ bigquery/tests/unit/test_table.py
@@ -12,14 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import email
-import io
-import json
 import unittest
 
 import mock
-from six.moves import http_client
-import pytest
 
 from google.cloud.bigquery.dataset import DatasetReference
 
@@ -1028,498 +1023,6 @@ class TestTable(unittest.TestCase, _SchemaBase):
         self.assertEqual(req['method'], 'POST')
         self.assertEqual(req['path'], '/%s' % PATH)
         self.assertEqual(req['data'], SENT)
-
-    def test__populate_view_use_legacy_sql_resource_w_existing_view(self):
-        query = 'select * from foo'
-        resource = {'view': {'query': query}}
-        client = mock.Mock(spec=[u'_credentials', '_http'])
-        client._http = mock.sentinel.http
-        dataset = DatasetReference(self.PROJECT, self.DS_ID)
-        table = self._make_one(dataset.table(self.TABLE_NAME), client=client)
-        table.view_use_legacy_sql = True
-
-        table._populate_view_use_legacy_sql_resource(resource)
-
-        self.assertEqual(
-            resource['view']['useLegacySql'], table.view_use_legacy_sql)
-        self.assertEqual(resource['view']['query'], query)
-
-    def test__get_transport(self):
-        client = mock.Mock(spec=[u'_credentials', '_http'])
-        client._http = mock.sentinel.http
-        dataset = DatasetReference(self.PROJECT, self.DS_ID)
-        table_ref = dataset.table(self.TABLE_NAME)
-        table = self._make_one(table_ref, client=client)
-
-        transport = table._get_transport(client)
-
-        self.assertIs(transport, mock.sentinel.http)
-
-    @staticmethod
-    def _mock_requests_response(status_code, headers, content=b''):
-        return mock.Mock(
-            content=content, headers=headers, status_code=status_code,
-            spec=['content', 'headers', 'status_code'])
-
-    def _mock_transport(self, status_code, headers, content=b''):
-        fake_transport = mock.Mock(spec=['request'])
-        fake_response = self._mock_requests_response(
-            status_code, headers, content=content)
-        fake_transport.request.return_value = fake_response
-        return fake_transport
-
-    def _initiate_resumable_upload_helper(self, num_retries=None):
-        from google.resumable_media.requests import ResumableUpload
-        from google.cloud.bigquery.table import _DEFAULT_CHUNKSIZE
-        from google.cloud.bigquery.table import _GENERIC_CONTENT_TYPE
-        from google.cloud.bigquery.table import _get_upload_headers
-        from google.cloud.bigquery.table import _get_upload_metadata
-
-        connection = _Connection()
-        client = _Client(self.PROJECT, connection=connection)
-        dataset = DatasetReference(self.PROJECT, self.DS_ID)
-        table_ref = dataset.table(self.TABLE_NAME)
-        table = self._make_one(table_ref, client=client)
-
-        # Create mocks to be checked for doing transport.
-        resumable_url = 'http://test.invalid?upload_id=hey-you'
-        response_headers = {'location': resumable_url}
-        fake_transport = self._mock_transport(
-            http_client.OK, response_headers)
-        client._http = fake_transport
-
-        # Create some mock arguments and call the method under test.
-        data = b'goodbye gudbi gootbee'
-        stream = io.BytesIO(data)
-        metadata = _get_upload_metadata(
-            'CSV', table._schema, table.project,
-            table.dataset_id, table.table_id)
-        upload, transport = table._initiate_resumable_upload(
-            client, stream, metadata, num_retries)
-
-        # Check the returned values.
-        self.assertIsInstance(upload, ResumableUpload)
-        upload_url = (
-            'https://www.googleapis.com/upload/bigquery/v2/projects/' +
-            self.PROJECT +
-            '/jobs?uploadType=resumable')
-        self.assertEqual(upload.upload_url, upload_url)
-        expected_headers = _get_upload_headers(connection.USER_AGENT)
-        self.assertEqual(upload._headers, expected_headers)
-        self.assertFalse(upload.finished)
-        self.assertEqual(upload._chunk_size, _DEFAULT_CHUNKSIZE)
-        self.assertIs(upload._stream, stream)
-        self.assertIsNone(upload._total_bytes)
-        self.assertEqual(upload._content_type, _GENERIC_CONTENT_TYPE)
-        self.assertEqual(upload.resumable_url, resumable_url)
-
-        retry_strategy = upload._retry_strategy
-        self.assertEqual(retry_strategy.max_sleep, 64.0)
-        if num_retries is None:
-            self.assertEqual(retry_strategy.max_cumulative_retry, 600.0)
-            self.assertIsNone(retry_strategy.max_retries)
-        else:
-            self.assertIsNone(retry_strategy.max_cumulative_retry)
-            self.assertEqual(retry_strategy.max_retries, num_retries)
-        self.assertIs(transport, fake_transport)
-        # Make sure we never read from the stream.
-        self.assertEqual(stream.tell(), 0)
-
-        # Check the mocks.
-        request_headers = expected_headers.copy()
-        request_headers['x-upload-content-type'] = _GENERIC_CONTENT_TYPE
-        fake_transport.request.assert_called_once_with(
-            'POST',
-            upload_url,
-            data=json.dumps(metadata).encode('utf-8'),
-            headers=request_headers,
-        )
-
-    def test__initiate_resumable_upload(self):
-        self._initiate_resumable_upload_helper()
-
-    def test__initiate_resumable_upload_with_retry(self):
-        self._initiate_resumable_upload_helper(num_retries=11)
-
-    def _do_multipart_upload_success_helper(
-            self, get_boundary, num_retries=None):
-        from google.cloud.bigquery.table import _get_upload_headers
-        from google.cloud.bigquery.table import _get_upload_metadata
-
-        connection = _Connection()
-        client = _Client(self.PROJECT, connection=connection)
-        dataset = DatasetReference(self.PROJECT, self.DS_ID)
-        table_ref = dataset.table(self.TABLE_NAME)
-        table = self._make_one(table_ref, client=client)
-
-        # Create mocks to be checked for doing transport.
-        fake_transport = self._mock_transport(http_client.OK, {})
-        client._http = fake_transport
-
-        # Create some mock arguments.
-        data = b'Bzzzz-zap \x00\x01\xf4'
-        stream = io.BytesIO(data)
-        metadata = _get_upload_metadata(
-            'CSV', table._schema, table.project,
-            table.dataset_id, table.table_id)
-        size = len(data)
-        response = table._do_multipart_upload(
-            client, stream, metadata, size, num_retries)
-
-        # Check the mocks and the returned value.
-        self.assertIs(response, fake_transport.request.return_value)
-        self.assertEqual(stream.tell(), size)
-        get_boundary.assert_called_once_with()
-
-        upload_url = (
-            'https://www.googleapis.com/upload/bigquery/v2/projects/' +
-            self.PROJECT +
-            '/jobs?uploadType=multipart')
-        payload = (
-            b'--==0==\r\n' +
-            b'content-type: application/json; charset=UTF-8\r\n\r\n' +
-            json.dumps(metadata).encode('utf-8') + b'\r\n' +
-            b'--==0==\r\n' +
-            b'content-type: */*\r\n\r\n' +
-            data + b'\r\n' +
-            b'--==0==--')
-        headers = _get_upload_headers(connection.USER_AGENT)
-        headers['content-type'] = b'multipart/related; boundary="==0=="'
-        fake_transport.request.assert_called_once_with(
-            'POST',
-            upload_url,
-            data=payload,
-            headers=headers,
-        )
-
-    @mock.patch(u'google.resumable_media._upload.get_boundary',
-                return_value=b'==0==')
-    def test__do_multipart_upload(self, get_boundary):
-        self._do_multipart_upload_success_helper(get_boundary)
-
-    @mock.patch(u'google.resumable_media._upload.get_boundary',
-                return_value=b'==0==')
-    def test__do_multipart_upload_with_retry(self, get_boundary):
-        self._do_multipart_upload_success_helper(get_boundary, num_retries=8)
-
-
-class TestTableUpload(object):
-    # NOTE: This is a "partner" to `TestTable` meant to test some of the
-    #       "upload" portions of `Table`. It also uses `pytest`-style tests
-    #       rather than `unittest`-style.
-
-    @staticmethod
-    def _make_table(transport=None):
-        from google.cloud.bigquery import _http
-        from google.cloud.bigquery import client
-        from google.cloud.bigquery import dataset
-        from google.cloud.bigquery import table
-
-        connection = mock.create_autospec(_http.Connection, instance=True)
-        client = mock.create_autospec(client.Client, instance=True)
-        client._connection = connection
-        client._credentials = mock.sentinel.credentials
-        client._http = transport
-        client.project = 'project_id'
-
-        dataset_ref = dataset.DatasetReference('project_id', 'test_dataset')
-        table_ref = dataset_ref.table('test_table')
-        table = table.Table(table_ref, client=client)
-
-        return table
-
-    @staticmethod
-    def _make_response(status_code, content='', headers={}):
-        """Make a mock HTTP response."""
-        import requests
-        response = requests.Response()
-        response.request = requests.Request(
-            'POST', 'http://example.com').prepare()
-        response._content = content.encode('utf-8')
-        response.headers.update(headers)
-        response.status_code = status_code
-        return response
-
-    @classmethod
-    def _make_do_upload_patch(cls, table, method, side_effect=None):
-        """Patches the low-level upload helpers."""
-        if side_effect is None:
-            side_effect = [cls._make_response(
-                http_client.OK,
-                json.dumps({}),
-                {'Content-Type': 'application/json'})]
-        return mock.patch.object(
-            table, method, side_effect=side_effect, autospec=True)
-
-    EXPECTED_CONFIGURATION = {
-        'configuration': {
-            'load': {
-                'sourceFormat': 'CSV',
-                'destinationTable': {
-                    'projectId': 'project_id',
-                    'datasetId': 'test_dataset',
-                    'tableId': 'test_table'
-                }
-            }
-        }
-    }
-
-    @staticmethod
-    def _make_file_obj():
-        return io.BytesIO(b'hello, is it me you\'re looking for?')
-
-    # High-level tests
-
-    def test_upload_from_file_resumable(self):
-        import google.cloud.bigquery.table
-
-        table = self._make_table()
-        file_obj = self._make_file_obj()
-
-        do_upload_patch = self._make_do_upload_patch(
-            table, '_do_resumable_upload')
-        with do_upload_patch as do_upload:
-            table.upload_from_file(file_obj, source_format='CSV')
-
-        do_upload.assert_called_once_with(
-            table._client,
-            file_obj,
-            self.EXPECTED_CONFIGURATION,
-            google.cloud.bigquery.table._DEFAULT_NUM_RETRIES)
-
-    def test_upload_file_resumable_metadata(self):
-        table = self._make_table()
-        file_obj = self._make_file_obj()
-
-        config_args = {
-            'source_format': 'CSV',
-            'allow_jagged_rows': False,
-            'allow_quoted_newlines': False,
-            'create_disposition': 'CREATE_IF_NEEDED',
-            'encoding': 'utf8',
-            'field_delimiter': ',',
-            'ignore_unknown_values': False,
-            'max_bad_records': 0,
-            'quote_character': '"',
-            'skip_leading_rows': 1,
-            'write_disposition': 'WRITE_APPEND',
-            'job_name': 'oddjob',
-            'null_marker': r'\N',
-        }
-
-        expected_config = {
-            'configuration': {
-                'load': {
-                    'sourceFormat': config_args['source_format'],
-                    'destinationTable': {
-                        'projectId': table.project,
-                        'datasetId': table.dataset_id,
-                        'tableId': table.table_id,
-                    },
-                    'allowJaggedRows': config_args['allow_jagged_rows'],
-                    'allowQuotedNewlines':
-                        config_args['allow_quoted_newlines'],
-                    'createDisposition': config_args['create_disposition'],
-                    'encoding': config_args['encoding'],
-                    'fieldDelimiter': config_args['field_delimiter'],
-                    'ignoreUnknownValues':
-                        config_args['ignore_unknown_values'],
-                    'maxBadRecords': config_args['max_bad_records'],
-                    'quote': config_args['quote_character'],
-                    'skipLeadingRows': config_args['skip_leading_rows'],
-                    'writeDisposition': config_args['write_disposition'],
-                    'jobReference': {'jobId': config_args['job_name']},
-                    'nullMarker': config_args['null_marker'],
-                },
-            },
-        }
-
-        do_upload_patch = self._make_do_upload_patch(
-            table, '_do_resumable_upload')
-        with do_upload_patch as do_upload:
-            table.upload_from_file(
-                file_obj, **config_args)
-
-        do_upload.assert_called_once_with(
-            table._client,
-            file_obj,
-            expected_config,
-            mock.ANY)
-
-    def test_upload_from_file_multipart(self):
-        import google.cloud.bigquery.table
-
-        table = self._make_table()
-        file_obj = self._make_file_obj()
-        file_obj_size = 10
-
-        do_upload_patch = self._make_do_upload_patch(
-            table, '_do_multipart_upload')
-        with do_upload_patch as do_upload:
-            table.upload_from_file(
-                file_obj, source_format='CSV', size=file_obj_size)
-
-        do_upload.assert_called_once_with(
-            table._client,
-            file_obj,
-            self.EXPECTED_CONFIGURATION,
-            file_obj_size,
-            google.cloud.bigquery.table._DEFAULT_NUM_RETRIES)
-
-    def test_upload_from_file_with_retries(self):
-        table = self._make_table()
-        file_obj = self._make_file_obj()
-        num_retries = 20
-
-        do_upload_patch = self._make_do_upload_patch(
-            table, '_do_resumable_upload')
-        with do_upload_patch as do_upload:
-            table.upload_from_file(
-                file_obj, source_format='CSV', num_retries=num_retries)
-
-        do_upload.assert_called_once_with(
-            table._client,
-            file_obj,
-            self.EXPECTED_CONFIGURATION,
-            num_retries)
-
-    def test_upload_from_file_with_rewind(self):
-        table = self._make_table()
-        file_obj = self._make_file_obj()
-        file_obj.seek(2)
-
-        with self._make_do_upload_patch(table, '_do_resumable_upload'):
-            table.upload_from_file(
-                file_obj, source_format='CSV', rewind=True)
-
-        assert file_obj.tell() == 0
-
-    def test_upload_from_file_failure(self):
-        from google.resumable_media import InvalidResponse
-        from google.cloud import exceptions
-
-        table = self._make_table()
-        file_obj = self._make_file_obj()
-
-        response = self._make_response(
-            content='Someone is already in this spot.',
-            status_code=http_client.CONFLICT)
-
-        do_upload_patch = self._make_do_upload_patch(
-            table, '_do_resumable_upload',
-            side_effect=InvalidResponse(response))
-
-        with do_upload_patch, pytest.raises(exceptions.Conflict) as exc_info:
-            table.upload_from_file(
-                file_obj, source_format='CSV', rewind=True)
-
-        assert response.text in exc_info.value.message
-        assert exc_info.value.errors == []
-
-    def test_upload_from_file_bad_mode(self):
-        table = self._make_table()
-        file_obj = mock.Mock(spec=['mode'])
-        file_obj.mode = 'x'
-
-        with pytest.raises(ValueError):
-            table.upload_from_file(
-                file_obj, source_format='CSV',)
-
-    # Low-level tests
-
-    @classmethod
-    def _make_resumable_upload_responses(cls, size):
-        """Make a series of responses for a successful resumable upload."""
-        from google import resumable_media
-
-        resumable_url = 'http://test.invalid?upload_id=and-then-there-was-1'
-        initial_response = cls._make_response(
-            http_client.OK, '', {'location': resumable_url})
-        data_response = cls._make_response(
-            resumable_media.PERMANENT_REDIRECT,
-            '', {'range': 'bytes=0-{:d}'.format(size - 1)})
-        final_response = cls._make_response(
-            http_client.OK,
-            json.dumps({'size': size}),
-            {'Content-Type': 'application/json'})
-        return [initial_response, data_response, final_response]
-
-    @staticmethod
-    def _make_transport(responses=None):
-        import google.auth.transport.requests
-
-        transport = mock.create_autospec(
-            google.auth.transport.requests.AuthorizedSession, instance=True)
-        transport.request.side_effect = responses
-        return transport
-
-    def test__do_resumable_upload(self):
-        file_obj = self._make_file_obj()
-        file_obj_len = len(file_obj.getvalue())
-        transport = self._make_transport(
-            self._make_resumable_upload_responses(file_obj_len))
-        table = self._make_table(transport)
-
-        result = table._do_resumable_upload(
-            table._client,
-            file_obj,
-            self.EXPECTED_CONFIGURATION,
-            None)
-
-        content = result.content.decode('utf-8')
-        assert json.loads(content) == {'size': file_obj_len}
-
-        # Verify that configuration data was passed in with the initial
-        # request.
-        transport.request.assert_any_call(
-            'POST',
-            mock.ANY,
-            data=json.dumps(self.EXPECTED_CONFIGURATION).encode('utf-8'),
-            headers=mock.ANY)
-
-    def test__do_multipart_upload(self):
-        transport = self._make_transport([self._make_response(http_client.OK)])
-        table = self._make_table(transport)
-        file_obj = self._make_file_obj()
-        file_obj_len = len(file_obj.getvalue())
-
-        table._do_multipart_upload(
-            table._client,
-            file_obj,
-            self.EXPECTED_CONFIGURATION,
-            file_obj_len,
-            None)
-
-        # Verify that configuration data was passed in with the initial
-        # request.
-        request_args = transport.request.mock_calls[0][2]
-        request_data = request_args['data'].decode('utf-8')
-        request_headers = request_args['headers']
-
-        request_content = email.message_from_string(
-            'Content-Type: {}\r\n{}'.format(
-                request_headers['content-type'].decode('utf-8'),
-                request_data))
-
-        # There should be two payloads: the configuration and the binary daya.
-        configuration_data = request_content.get_payload(0).get_payload()
-        binary_data = request_content.get_payload(1).get_payload()
-
-        assert json.loads(configuration_data) == self.EXPECTED_CONFIGURATION
-        assert binary_data.encode('utf-8') == file_obj.getvalue()
-
-    def test__do_multipart_upload_wrong_size(self):
-        table = self._make_table()
-        file_obj = self._make_file_obj()
-        file_obj_len = len(file_obj.getvalue())
-
-        with pytest.raises(ValueError):
-            table._do_multipart_upload(
-                table._client,
-                file_obj,
-                {},
-                file_obj_len+1,
-                None)
 
 
 class Test_parse_schema_resource(unittest.TestCase, _SchemaBase):
